% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_annuity_functions.R
\name{annuity_due}
\alias{annuity_due}
\title{Annuity-Due}
\usage{
annuity_due(rate, n)
}
\arguments{
\item{rate}{The interest rate per period.}

\item{n}{The number of periods.}
}
\value{
The present value of an annuity-due.
}
\description{
Annuity-Due
}
\examples{
annuity_due(0.05, 10)
}
