% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_annuity_functions.R
\name{continuous_annuity}
\alias{continuous_annuity}
\title{Continuous Annuity}
\usage{
continuous_annuity(rate, n)
}
\arguments{
\item{rate}{The interest rate per period.}

\item{n}{The number of periods.}
}
\value{
The present value of a continuous annuity.
}
\description{
Continuous Annuity
}
\examples{
continuous_annuity(0.05, 10)
}
