% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_of_value.R
\name{equation_of_value}
\alias{equation_of_value}
\title{Equation of Value}
\usage{
equation_of_value(cash_flows, times, rate)
}
\arguments{
\item{cash_flows}{A vector of cash flows.}

\item{times}{A vector of times at which the cash flows occur.}

\item{rate}{The interest rate per period.}
}
\value{
The present value of the cash flows.
}
\description{
Equation of Value
}
\examples{
equation_of_value(c(1000, 2000, 3000), c(1, 2, 3), 0.05)
}
