% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_annuity_functions.R
\name{immediate_annuity}
\alias{immediate_annuity}
\title{Immediate Annuity}
\usage{
immediate_annuity(rate, n)
}
\arguments{
\item{rate}{The interest rate per period.}

\item{n}{The number of periods.}
}
\value{
The present value of an immediate annuity.
}
\description{
Immediate Annuity
}
\examples{
immediate_annuity(0.05, 10)
}
