% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_level_annuity_functions.R
\name{non_level_annuity_due}
\alias{non_level_annuity_due}
\title{Non-Level Annuities-Due with Varying Payments}
\usage{
non_level_annuity_due(payments, rate)
}
\arguments{
\item{payments}{A vector of payments.}

\item{rate}{The interest rate per period.}
}
\value{
The present value of non-level annuities-due.
}
\description{
Non-Level Annuities-Due with Varying Payments
}
\examples{
non_level_annuity_due(c(100, 200, 300), 0.05)
}
