% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_measures.R
\name{simple_interest}
\alias{simple_interest}
\title{Simple Interest Calculation}
\usage{
simple_interest(principal, rate, time)
}
\arguments{
\item{principal}{The initial amount of money.}

\item{rate}{The interest rate per period.}

\item{time}{The number of periods the money is invested or borrowed for.}
}
\value{
The interest accrued after \code{time} periods.
}
\description{
Simple Interest Calculation
}
\examples{
simple_interest(1000, 0.05, 10)
}
